requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
PyQt5>=5.15.0
Pillow>=9.0.0
